refactor(ui): simplify location table
