+++ docs/FALLOW_REMEDIATION_PLAN.md
@@ -77,7 +77,7 @@ Run browser, API, or migration tests for each affected slice.
 
 Purpose: close the residual first-party findings through behavior-preserving refactors and verified cleanup.
 
-- [ ] Trace and reduce the public export surface in `src/stores/playthroughs/` without removing runtime or external consumers.
+- [x] Trace and reduce the public export surface in `src/stores/playthroughs/` without removing runtime or external consumers.
 - [ ] Remove remaining verified unused files, exports, and types in coherent module batches.
 - [ ] Add behavior and accessibility coverage for context-menu, summary-card, PC, team, and location interaction paths.
 - [ ] Refactor context-menu action construction and shared behavior before extracting common code.
