+++ src/components/PokemonCombobox/PokemonCombobox.tsx
@@ -30,7 +30,7 @@ import {
   useAllPokemon,
   usePokemonSearch,
 } from "@/loaders/pokemon";
-import { useEncounters, useGameMode } from "@/stores/playthroughs";
+import { useEncounters, useGameMode } from "@/stores/playthroughs/hooks";
 import type { EncounterSource } from "@/types/encounters";
 import { buildCapturedSpeciesIdSet } from "@/utils/encounter-utils";
 import { DraggableComboboxSprite } from "./DraggableComboboxSprite";
