--- src/stores/playthroughs.ts
@@ -1,2 +1,0 @@
-// Re-export everything from the new modular structure for backward compatibility
-export * from "./playthroughs/index";
